requirements.txt
Flask==2.2.2
flask-sqlalchemy==3.0.2
Flask-WTF==1.0.1
Flask-Bcrypt==1.0.1
mysql-connector-python==8.0.31
